src/modules/users/useCases/createUser/CreateUserUseCase.ts
import { inject, injectable } from 'tsyringe';

import { AppError } from '@shared/errors/AppError';
import { IHashProvider } from '@shared/providers/HashProvider/IHashProvider';

import { User } from '../../infra/typeorm/models/User';
import { IUsersRepository } from '../../repositories/IUsersRepository';

type ICreateUserRequest = {
  name: string;
  email: string;
  password: string;
};

@injectable()
export class CreateUserUseCase {
  constructor(
    @inject('UsersRepository')
    private readonly usersRepository: IUsersRepository,
    @inject('HashProvider')
    private readonly hashProvider: IHashProvider,
  ) {}

  async execute({ name, email, password }: ICreateUserRequest): Promise<User> {
    const userAlreadyExists = await this.usersRepository.findByEmail(email);

    if (userAlreadyExists) {
      throw new AppError('User already exists');
    }

    const passwordHash = await this.hashProvider.generateHash(password);

    const user = await this.usersRepository.create({
      name,
      email,
      password: passwordHash,
    });

    return user;
  }
}
